Put the namespace Detail into a separate file
