+++ ir/CMakeLists.txt
@@ -45,6 +45,7 @@ set (IR_HDRS
   ir-tree-macros.h
   ir.h
   ir-traversal.h
+  ir-traversal-internal.h
   irutils.h
   json_generator.h
   json_loader.h
