+++ ir/ir-traversal.h
@@ -23,9 +23,9 @@ limitations under the License.
 
 #ifndef IR_IR_TRAVERSAL_H_
 #define IR_IR_TRAVERSAL_H_
+#define IR_IR_TRAVERSAL_INTERNAL_ENABLE
 
-#include <lib/exceptions.h>
-#include <lib/rtti_utils.h>
+#include <cstddef>
 
 namespace P4::IR::Traversal {
 
@@ -52,80 +52,12 @@ struct Index {
     size_t value;
 };
 
-namespace Detail {
+}  // namespace P4::IR::Traversal
 
-/// @brief Internal, don't use directly.
-/// The class exists so that the functions can be mutually recursive.
-struct Traverse {
-    template <typename Obj, typename T>
-    static const Obj *modify(const Obj *, Assign<const T *> asgn) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        return asgn.value;
-    }
+// Put the internals out of the main file to increase its readability.
+#include "ir-traversal-internal.h"
 
-    template <typename Obj, typename T>
-    static const Obj *modify(Obj *, Assign<const T *> asgn) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        return asgn.value;
-    }
-
-    template <typename Obj, typename T>
-    static Obj *modify(Obj *obj, Assign<T> &&asgn) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        *obj = std::move(asgn.value);
-        return obj;
-    }
-
-    template <typename Obj, typename T>
-    static Obj *modify(Obj *obj, const Assign<T> &asgn) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        *obj = asgn.value;
-        return obj;
-    }
-
-    template <typename Obj, typename Fn>
-    static decltype(auto) modify(Obj *obj, Fn fn) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        return fn(obj);
-    }
-
-    template <typename Obj, typename... Selectors>
-    static Obj *modify(Obj *obj, Index idx, Selectors &&...selectors) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        BUG_CHECK(obj->size() > idx.value, "Index %1% out of bounds of %2%", idx.value, obj);
-        modifyRef((*obj)[idx.value], std::forward<Selectors>(selectors)...);
-        return obj;
-    }
-
-    template <typename Obj, typename To, typename... Selectors>
-    static To *modify(Obj *obj, RTTI::Detail::ToType<To> cast, Selectors &&...selectors) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        auto *casted = cast(obj);
-        BUG_CHECK(casted, "Cast of %1% failed", obj);
-        return modify(casted, std::forward<Selectors>(selectors)...);
-    }
-
-    template <typename Obj, typename T, typename Sub, typename... Selectors>
-    static Obj *modify(Obj *obj, Sub T::*member, Selectors &&...selectors) {
-        static_assert(!std::is_const_v<Obj>, "Cannot modify constant object");
-        modifyRef(obj->*member, std::forward<Selectors>(selectors)...);
-        return obj;
-    }
-
-    template <typename T, typename... Selectors>
-    static void modifyRef(T &ref, Selectors &&...selectors) {
-        if constexpr (std::is_pointer_v<T>) {
-            ref = modify(ref->clone(), std::forward<Selectors>(selectors)...);
-        } else {
-            auto *res = modify(&ref, std::forward<Selectors>(selectors)...);
-            if (&ref != res) {
-                ref = *res;
-            }
-        }
-    }
-};
-
-}  // namespace Detail
+namespace P4::IR::Traversal {
 
 /// @brief Given an object @p obj and a series of selector (ending with a modifier), modify the @p
 /// obj's sub object at the selected path. %This is useful for deeper modification of objects that
@@ -199,4 +131,5 @@ Obj *apply(const Obj *obj, Selectors &&...selectors) {
 
 }  // namespace P4::IR::Traversal
 
+#undef IR_IR_TRAVERSAL_INTERNAL_ENABLE
 #endif  // IR_IR_TRAVERSAL_H_
